LayoutTests/platform/efl/inspector-protocol/debugger/setBreakpoint-dfg-and-modify-local-expected.txt

Debugger.evaluateOnCallFrame in a DFG compiled functions.

Found breakpoint.js
result: 502500
dfg function warmed up

Breakpoint set in breakpointBasic()
Hit Breakpoint!
Evaluating in DFG frame at frame[1]: 'x = 20000;'
inside breakpointBasic
PASS: result is 20003
Resumed from breakpoint
Test complete
